Allow for multilingual FX Help tooltips
This PR allows for showing the FX Help tooltips in the user's preferred language, as indicated in" Preferences". If a folder by the name of the language string exists in "$stuff/doc" (e.g. "$stuff/doc/English"), the help files there will be shown first. If the folder or requested file cannot be found, the files in "$stuff/doc" will be shown by default. This is my first code PR, and I am not a programmer, so please be lenient. Thanks to @RodneyBaker for coming up with the idea, to @turtletooth and @shun-iwasawa for helping me resolve my build issues, and finally to @manongjohn for providing encouragement and the logic behind these changes!